Feat:"Update Authorization"

+++ Controllers/contactController.js
@@ -4,7 +4,7 @@ const Contacts = require("../Models/contactModels")
 
 //GET ALL CONTACTS
 const getAllContacts = asyncHandler(async (req,res)=>{
-    const contacts = await Contacts.find()
+    const contacts = await Contacts.find({userId:req.user.id})
     res.status(200).json(contacts)
 })
 
@@ -30,7 +30,8 @@ const createNewContact = asyncHandler(async(req,res)=>{
     const contact = await Contacts.create({
         name,
         email,
-        phone
+        phone,
+        userId: req.user.id
     })
     //OR ALTERNATIVE
     // const contact = await Contacts.create(Name:req.body.Name,Email:req.body.Email,Phone:req.body.Phone)
@@ -44,6 +45,10 @@ const updateContact = asyncHandler(async(req,res)=>{
         res.status(404)
         throw new Error("Contact not found")
     }
+    if(contact.userId.toString() !== req.user.id){
+        res.status(403)
+        throw new Error("User not authorized to update this contact")
+    }
     // {new:true} will return the updated document after the update operation
     const updatedContact = await Contacts.findByIdAndUpdate(req.params.id,req.body,{new:true})
     res.status(200).json(updatedContact)
@@ -54,6 +59,10 @@ const deleteContact = asyncHandler(async(req,res)=>{
     if(!contact){
         res.status(404)
         throw new Error("Contact not Found")
+    }
+    if(contact.userId.toString() !== req.user.id){
+        res.status(403)
+        throw new Error("User not authorized to delete this contact")
     }
     await Contacts.findByIdAndDelete(req.params.id)
     res.status(200).json(contact)
